File override now only if the file doesn't exists

+++ reformcloud-global/reformcloud-client/src/main/java/systems/reformcloud/serverprocess/startup/CloudServerStartupHandler.java
@@ -179,8 +179,11 @@ public final class CloudServerStartupHandler implements Serializable {
 
         this.processStartupStage = ProcessStartupStage.PREPARING;
         if (serverStartupInfo.getServerGroup().getSpigotVersions().equals(SpigotVersions.SHORTSPIGOT_1_12_2)) {
-            FileUtils.copyCompiledFile("reformcloud/spigot/spigot.yml", path + "/configs/spigot.yml");
-            FileUtils.copyCompiledFile("reformcloud/default/server.properties", path + "/configs/server.properties");
+            if (!Files.exists(Paths.get(path + "/configs/spigot.yml")))
+                FileUtils.copyCompiledFile("reformcloud/spigot/spigot.yml", path + "/configs/spigot.yml");
+
+            if (!Files.exists(Paths.get(path + "/configs/server.properties")))
+                FileUtils.copyCompiledFile("reformcloud/default/server.properties", path + "/configs/server.properties");
         } else {
             if (serverStartupInfo.getServerGroup().getSpigotVersions().equals(SpigotVersions.GLOWSTONE_1_12_2)) {
                 if (!Files.exists(Paths.get(path + "/config")))
@@ -188,8 +191,11 @@ public final class CloudServerStartupHandler implements Serializable {
                 if (!Files.exists(Paths.get(path + "/config/glowstone.yml")))
                     FileUtils.copyCompiledFile("reformcloud/glowstone/glowstone.yml", path + "/config/glowstone.yml");
             } else {
-                FileUtils.copyCompiledFile("reformcloud/spigot/spigot.yml", path + "/spigot.yml");
-                FileUtils.copyCompiledFile("reformcloud/default/server.properties", path + "/server.properties");
+                if (!Files.exists(Paths.get(path + "/spigot.yml")))
+                    FileUtils.copyCompiledFile("reformcloud/spigot/spigot.yml", path + "/spigot.yml");
+
+                if (!Files.exists(Paths.get(path + "/server.properties")))
+                    FileUtils.copyCompiledFile("reformcloud/default/server.properties", path + "/server.properties");
             }
         }
